DOWNGRADE to NETframework
I Initially thought about doing it with NETframework but suspected its datedness to cause limitations (considering that winforms is already dated), which is somewhat true after recent testing. so far user-wise speaking, the only limitation I've encountered is delayed drawing on window resize, which I've already implemented a sort of fix to. On the bonus side, the file size is now drastically lower, memory usage much more improved/idealized. I'd probably be rewriting this project on a new framework eitherways because this is too goofy.